refactor: move test files out of __tests__ folder
